Added Zach's slides to the main file.
I added Zach’s slides to the template so they are all in one file now.
I fixed up the headers for the slides so that we are being consistent in their use.  i.e. for each slide we have the model type at the top then the subheading.  I don’t think it’s necessary to put Non-Linear Regression at the top of every slide, so I removed that.
I also fixed the css for the code blocks so they look A LOT better now.
I liked the ioslides format (that Zach was using, but if your slide was too tall it cut off the bottom with no scroll bar, so we can’t use it unless everyone fixes up their slides to make sure they don’t run off the bottom.  It also left a big black border around the edges that I couldn’t get rid of.
I also ran a spell check on everything and fixed a few typos.
I still have to add my slides!  But it looks like one coherent presentation now instead of two separate presentations.